+++ packages/cubejs-query-orchestrator/src/orchestrator/RedisPool.ts
@@ -9,10 +9,15 @@ export interface RedisPoolOptions {
   createClient?: CreateRedisClientFn;
   destroyClient?: (client: Redis.Redis) => PromiseLike<void>;
 }
+
+const MAX_ALLOWED_POOL_ERRORS = 100;
+
 export class RedisPool {
   protected readonly pool: Pool<Redis.Redis>|null = null;
 
   protected readonly create: CreateRedisClientFn|null = null;
+  
+  protected poolErrors: number = 0;
 
   public constructor(options: RedisPoolOptions = {}) {
     const defaultMin = process.env.CUBEJS_REDIS_POOL_MIN ? parseInt(process.env.CUBEJS_REDIS_POOL_MIN, 10) : 2;
@@ -27,13 +32,27 @@ export class RedisPool {
       idleTimeoutMillis: 5000,
       evictionRunIntervalMillis: 5000
     };
-
+    
     const create = options.createClient || (async () => createRedisClient(process.env.REDIS_URL));
 
     if (max > 0) {
       const destroy = options.destroyClient || (async (client) => client.disconnect());
 
+<<<<<<< HEAD
       this.pool = genericPool.createPool<Redis.Redis>({ create, destroy }, opts);
+=======
+      this.pool = genericPool.createPool<AsyncRedisClient>({ create, destroy }, opts);
+      
+      this.pool.on('factoryCreateError', (error) => {
+        this.poolErrors++;
+        // prevent the infinite loop when pool creation fails too many times
+        if (this.poolErrors > MAX_ALLOWED_POOL_ERRORS) {
+          // @ts-ignore
+          // eslint-disable-next-line
+          this.pool._waitingClientsQueue.dequeue().reject(error);
+        }
+      });
+>>>>>>> master
     } else {
       // fallback to un-pooled behavior if pool max is 0
       this.create = create;