+++ packages/cubejs-schema-compiler/src/compiler/DynamicReference.js
@@ -1,0 +1,6 @@
+export class DynamicReference {
+  constructor(memberNames, fn) {
+    this.memberNames = memberNames;
+    this.fn = fn;
+  }
+}