+++ packages/cubejs-schema-compiler/src/compiler/UserError.js
@@ -1,0 +1,6 @@
+export class UserError extends Error {
+  constructor(message) {
+    super(message);
+    this.type = 'UserError';
+  }
+}